Move semaphore from asr to microphone node
